+++ Editor/Inspector/MaterialSwap/MatSwapEditor.cs
@@ -99,9 +99,13 @@ namespace nadena.dev.modular_avatar.core.editor
                     });
                 }
 #if UNITY_6000_3_OR_NEWER
-                menu.DropDown(_fromField.worldBound, _fromField,DropdownMenuSizeMode.Auto);
+                menu.DropDown(_fromField.worldBound, _fromField, DropdownMenuSizeMode.Auto);
+#else
+#if UNITY_6000_0_OR_NEWER
+                menu.DropDown(_fromField.worldBound, _fromField, false, false);
 #else
                 menu.DropDown(_fromField.worldBound, _fromField);
+#endif
 #endif
             };
         }