+++ CHANGELOG.md
@@ -15,6 +15,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - [#1987] Added support for remapping blendshapes using a curve in `MA Blendshape Sync`
 
 ### Fixed
+- [#2007] Improved compatibility with Unity 6000.0
 - [#1947] Fixed an issue where the scale adjuster preview would not update properly when objects move a small distance.
 - [#1994] Improved performance of `Mesh Cutter` previews and builds
 - [#1922] Fixed an issue where the avatar's original first FX layer would be disabled in MMD worlds when reactive